refactor: split models file out
