+++ chuni_penguin/database/__init__.py
@@ -1,21 +1,11 @@
-from .models import (
-    Alias,
-    Chart,
-    Cookie,
-    Course,
-    CourseTrack,
-    Denylist,
-    EasterEggFound,
-    GuessScore,
-    PendingKamaitachiImport,
-    PersonalBest,
-    Prefix,
-    SdvxinChartView,
-    Song,
-    SongJacket,
-    UserConfig,
-    course_track_charts,
-)
+from .bot import Denylist
+from .courses import Course, CourseTrack, course_track_charts
+from .guilds import Prefix
+from .kamaitachi import PendingKamaitachiImport
+from .minigames import GuessScore
+from .pbs import PersonalBest
+from .songs import Alias, Chart, SdvxinChartView, Song, SongJacket
+from .users import Cookie, EasterEggFound, UserConfig
 
 __all__ = (
     "Alias",
