+++ chuni_penguin/database/alembic/env.py
@@ -3,7 +3,7 @@ from logging.config import fileConfig
 from alembic import context
 from sqlalchemy import engine_from_config, pool
 
-from chuni_penguin.database.models import Base
+from chuni_penguin.database.base import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
