+++ chuni_penguin/database/base.py
@@ -1,0 +1,45 @@
+from datetime import datetime, timezone
+
+from sqlalchemy import DateTime, Dialect, TypeDecorator
+from sqlalchemy.ext.asyncio import AsyncAttrs
+from sqlalchemy.orm import DeclarativeBase
+
+
+class DateTimeUTC(TypeDecorator[datetime]):
+    """Timezone Aware DateTime.
+
+    Ensure UTC is stored in the database and that TZ aware dates are returned for all dialects.
+    """
+
+    impl = DateTime(timezone=True)
+    cache_ok = True
+
+    @property
+    def python_type(self) -> type[datetime]:
+        return datetime
+
+    def process_bind_param(
+        self, value: datetime | None, dialect: Dialect
+    ) -> datetime | None:
+        if value is None:
+            return value
+        if not value.tzinfo:
+            msg = "tzinfo is required"
+            raise TypeError(msg)
+        return value.astimezone(timezone.utc)
+
+    def process_result_value(
+        self, value: datetime | None, dialect: Dialect
+    ) -> datetime | None:
+        if value is None:
+            return value
+        if value.tzinfo is None:
+            return value.replace(tzinfo=timezone.utc)
+        return value
+
+
+class Base(DeclarativeBase, AsyncAttrs):
+    pass
+
+
+Base.registry.type_annotation_map[datetime] = DateTimeUTC
